Nest AudioAtom a level down from App to avoid App rerenders
